CFITSIO: multiple rowfilter and/or colfilter expressions are now allowed in on-the-fly file opening syntax.
Example:
  myfile.fits[col *][(#ROW < 10)][col X=Y+1][(X > 1.0)]
is tranlated to the equivalent form
  myfile.fits[col *;X=Y+1;][(#ROW < 10)&&(X > 1.0)]

Because of the behind the scenes way this is implemented, it
is not permitted to use the @filename.txt when multiple
expressions are used, either for column or row filtering.

In the previous version of CFITSIO any use of multiple expressions
would lead to an error, or at very least undefined or unpredictable
behavior.  This change normalizes the behavior and extends it
in an intuitive way.

It also permits tools like 'ftmerge' or 'ftstat', which rely upon
extended syntax filtering internally to accept filenames that have
extended syntax themselves.

I verified that this works with a few quick test cases; that it
correctly generates an error when @filename.txt is used along
with multiple expressions; and that old behavior is retained.

I updated the CFITSIO and FITSIO .tex file, as well as the rowfilter
and colfilter fhelp files.

--CM