No longer allow the SUBTRACTIVE_DITHER_2 quantization method to be used when writing an image using HCOMPRESS_1 compression.  If the user tries this, they will get SUBTRACTIVE_DITHER_1 instead.
From Bill: "subtractive_dither_2 is intended for the special case where
'bad pixels' in the image have been artificially set to zero to indicate
that they have no valid value.  In this case it is important to exactly
preserve the zero value during any image compression process.  This
option is not supported when using the lossy Hcompress method ..."