Added a series of new #define statements to generate an argument list for function prototypes when using PROTOCCALLSFFUNn.
Beginning with the C23 standard, gcc no longer compiles when the
function prototype argument list is left empty (unless the function
actually does take no arguments).  This is a partial fix which takes
care of all the cases except where the Fortran function is returning a
string (which we don't seem to be using in Heasoft).