Use NO_DITHER when ZZERO/ZSCALE present but ZQUANTIZ is not.
NO_DITHER being the default is clearly stated in 10.1.2 of the
standard.

It's not clear if quantize_method=0 and quantize_method=NO_DITHER are
actually being treated as identical elsewhere in the code, resulting in
correct behavior prior to this change, but this makes adherence to the
standard much clearer.